requirements.txt
llama-cpp-python
langchain
langchain-community
langchain-groq
langchain-text-splitters
numpy
pymongo
tiktoken
retry
requests
pydantic
feedparser
newspaper3k
tldextract
beautifulsoup4
lxml_html_clean
icecream
python-dotenv
persist-queue
chromadb
scikit-learn
pysqlite3-binary

